Fix Cyclic Import Error
This is Option #1

This fixes a pylint error that was introduced when I added the
version feature.  It caused a Cyclic Import error because I was
re-importing the base insteon_mqtt module over again.  While
things worked, I agree with pylint that this is a dangerous
design.

I am not sure what the proper solution is.  I would like to access
the __version__ variable in the base module from other classes
that do not inherit from this base module.

This solution uses __main__ which looks reasonable elegant enough
but requires ignoring a linter error that __main__ is undefined,
because it is, until runtime.

Not super thrilled by this as someone calling the start method
not from the command line could get an error.

Error introduced in: 9e9f7061c47b2174bcd0e9e8cbee21aa49d47eaa